Migrates project to Blazor Server
Converts the project from Blazor WebAssembly to Blazor Server.

This migration provides improved SEO, enhanced security, and
the ability to leverage server-side resources directly. It
also addresses the removal of the now obsolete Program.cs file,
replaces the App.razor file with Routes.razor, and updates the
project file.

Additionally, configures interactive server rendering for the Home
and MobileAppReport components and fixes a bug where the import
message was not being reset.